requirements.txt
fastapi
python-dotenv
langchain
langchain-community
langchain-openai
langchain-ollama
langchain-mongodb
langgraph
pymongo
uvicorn